requirements.txt
discord.py>=2.7.1,<3
aiosqlite>=0.22.1
python-dotenv>=1.0.0
matplotlib>=3.8
